Move to context. Cleanup
